correlated_stack: move TiffFrame and TiffStack to widefield
